поправки с предыдущего дз
